clean up documentation and notes
